ref: separate sale and redeem execute functions into different files
