+++ contracts/fungible-tokens/andromeda-cw20-exchange/src/contract.rs
@@ -1,41 +1,44 @@
 use andromeda_fungible_tokens::cw20_exchange::{
-    Cw20HookMsg, ExecuteMsg, InstantiateMsg, QueryMsg, Redeem, RedeemResponse, Sale,
-    SaleAssetsResponse, SaleResponse, TokenAddressResponse,
+    Cw20HookMsg, ExecuteMsg, InstantiateMsg, QueryMsg, RedeemResponse, SaleAssetsResponse,
+    SaleResponse, TokenAddressResponse,
 };
 use andromeda_std::{
     ado_base::{InstantiateMsg as BaseInstantiateMsg, MigrateMsg},
     ado_contract::ADOContract,
     amp::Recipient,
     andr_execute_fn,
-    common::{
-        context::ExecuteContext,
-        expiration::Expiry,
-        msg_generation::{generate_transfer_message, generate_transfer_message_recipient},
-        Milliseconds, MillisecondsDuration,
-    },
+    common::context::ExecuteContext,
     error::ContractError,
 };
 use cosmwasm_std::{
-    attr, ensure, entry_point, from_json, to_json_binary, Binary, Decimal, Deps, DepsMut, Env,
-    MessageInfo, Reply, Response, StdError, Uint128,
+    ensure, entry_point, from_json, to_json_binary, Binary, Deps, DepsMut, Env, MessageInfo, Reply,
+    Response, StdError,
 };
 use cw20::Cw20ReceiveMsg;
 use cw_asset::AssetInfo;
 use cw_storage_plus::Bound;
-use cw_utils::one_coin;
 
-use crate::state::{REDEEM, SALE, TOKEN_ADDRESS};
+use crate::{
+    execute_redeem::{
+        execute_cancel_redeem, execute_redeem, execute_redeem_native, execute_start_redeem,
+        execute_start_redeem_native,
+    },
+    execute_sale::{
+        execute_cancel_sale, execute_purchase, execute_purchase_native, execute_start_sale,
+    },
+    state::{REDEEM, SALE, TOKEN_ADDRESS},
+};
 
 // version info for migration info
 const CONTRACT_NAME: &str = "crates.io:andromeda-cw20-exchange";
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
 
 /// ID used for any refund sub messgaes
-const REFUND_REPLY_ID: u64 = 1;
+pub const REFUND_REPLY_ID: u64 = 1;
 /// ID used for any purchased token transfer sub messages
-const PURCHASE_REPLY_ID: u64 = 2;
+pub const PURCHASE_REPLY_ID: u64 = 2;
 /// ID used for transfer to sale recipient
-const RECIPIENT_REPLY_ID: u64 = 3;
+pub const RECIPIENT_REPLY_ID: u64 = 3;
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(
@@ -160,521 +163,6 @@ pub fn execute_receive(
             execute_redeem(ctx, amount_sent, asset_sent, recipient, &sender)
         }
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn execute_start_sale(
-    ctx: ExecuteContext,
-    amount: Uint128,
-    asset: AssetInfo,
-    exchange_rate: Uint128,
-    // The original sender of the CW20::Send message
-    sender: String,
-    // The recipient of the sale proceeds
-    recipient: Option<Recipient>,
-    start_time: Option<Expiry>,
-    duration: Option<MillisecondsDuration>,
-) -> Result<Response, ContractError> {
-    let recipient = Recipient::validate_or_default(recipient, &ctx, sender.as_str())?;
-
-    let ExecuteContext {
-        deps, env, info, ..
-    } = ctx;
-
-    let token_addr = TOKEN_ADDRESS
-        .load(deps.storage)?
-        .get_raw_address(&deps.as_ref())?;
-
-    ensure!(
-        asset != AssetInfo::Cw20(token_addr.clone()),
-        ContractError::InvalidAsset {
-            asset: asset.to_string()
-        }
-    );
-    ensure!(
-        !exchange_rate.is_zero(),
-        ContractError::InvalidZeroAmount {}
-    );
-    ensure!(
-        ctx.contract.is_contract_owner(deps.storage, &sender)?,
-        ContractError::Unauthorized {}
-    );
-    // Message sender in this case should be the token address
-    ensure!(
-        info.sender == token_addr,
-        ContractError::InvalidFunds {
-            msg: "Incorrect CW20 provided for sale".to_string()
-        }
-    );
-
-    let start_time = match start_time {
-        Some(s) => {
-            // Check that the start time is in the future
-            s.validate(&env.block)?
-        }
-        // Set start time to current time if not provided
-        None => Expiry::FromNow(Milliseconds::zero()),
-    }
-    .get_time(&env.block);
-
-    let end_time = match duration {
-        Some(e) => {
-            if e.is_zero() {
-                // If duration is 0, set end time to none
-                None
-            } else {
-                // Set end time to current time + duration
-                Some(Expiry::FromNow(e).get_time(&env.block))
-            }
-        }
-        None => None,
-    };
-
-    // Do not allow duplicate sales
-    let current_sale = SALE.may_load(deps.storage, &asset.to_string())?;
-    ensure!(current_sale.is_none(), ContractError::SaleNotEnded {});
-
-    let sale = Sale {
-        amount,
-        exchange_rate,
-        recipient,
-        start_time,
-        end_time,
-    };
-    SALE.save(deps.storage, &asset.to_string(), &sale)?;
-
-    Ok(Response::default().add_attributes(vec![
-        attr("action", "start_sale"),
-        attr("asset", asset.to_string()),
-        attr("rate", exchange_rate),
-        attr("amount", amount),
-        attr("start_time", start_time.to_string()),
-        attr("end_time", end_time.unwrap_or_default().to_string()),
-    ]))
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn execute_start_redeem(
-    ctx: ExecuteContext,
-    amount: Uint128,
-    asset: AssetInfo,
-    redeem_asset: AssetInfo,
-    exchange_rate: Decimal,
-    // The original sender of the CW20::Send message
-    sender: String,
-    // The recipient of the sale proceeds
-    recipient: Option<Recipient>,
-    start_time: Option<Expiry>,
-    duration: Option<MillisecondsDuration>,
-) -> Result<Response, ContractError> {
-    let recipient = Recipient::validate_or_default(recipient, &ctx, sender.as_str())?;
-
-    let ExecuteContext { deps, env, .. } = ctx;
-    // Ensure the redeem asset is not the token address, since we will be redeeming it
-    ensure!(
-        asset != redeem_asset,
-        ContractError::InvalidAsset {
-            asset: asset.to_string()
-        }
-    );
-    ensure!(
-        !exchange_rate.is_zero(),
-        ContractError::InvalidZeroAmount {}
-    );
-    ensure!(
-        ctx.contract.is_contract_owner(deps.storage, &sender)?,
-        ContractError::Unauthorized {}
-    );
-
-    let start_time = match start_time {
-        Some(s) => {
-            // Check that the start time is in the future
-            s.validate(&env.block)?
-        }
-        // Set start time to current time if not provided
-        None => Expiry::FromNow(Milliseconds::zero()),
-    }
-    .get_time(&env.block);
-
-    let end_time = match duration {
-        Some(e) => {
-            if e.is_zero() {
-                // If duration is 0, set end time to none
-                None
-            } else {
-                // Set end time to current time + duration
-                Some(Expiry::FromNow(e).get_time(&env.block))
-            }
-        }
-        None => None,
-    };
-
-    // Do not allow duplicate sales
-    let current_redeem = REDEEM.may_load(deps.storage, &redeem_asset.inner())?;
-    if let Some(redeem) = current_redeem {
-        // The old redeem should either be expired or have no amount left
-        ensure!(
-            redeem.start_time.is_expired(&env.block) || redeem.amount.is_zero(),
-            ContractError::RedeemNotEnded {}
-        );
-    }
-
-    let redeem = Redeem {
-        asset: asset.clone(),
-        amount,
-        amount_paid_out: Uint128::zero(),
-        exchange_rate,
-        recipient,
-        start_time,
-        end_time,
-    };
-    REDEEM.save(deps.storage, &redeem_asset.inner(), &redeem)?;
-
-    Ok(Response::default().add_attributes(vec![
-        attr("action", "start_redeem"),
-        attr("redeem_asset", redeem_asset.to_string()),
-        attr("asset", asset.to_string()),
-        attr("rate", exchange_rate.to_string()),
-        attr("amount", amount),
-        attr("start_time", start_time.to_string()),
-        attr("end_time", end_time.unwrap_or_default().to_string()),
-    ]))
-}
-
-pub fn execute_purchase(
-    ctx: ExecuteContext,
-    amount_sent: Uint128,
-    asset_sent: AssetInfo,
-    recipient: Recipient,
-    // For refund purposes
-    sender: &str,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { deps, .. } = ctx;
-    // deps.api.addr_validate(recipient)?;
-    let mut resp = Response::default();
-
-    let Some(mut sale) = SALE.may_load(deps.storage, &asset_sent.to_string())? else {
-        return Err(ContractError::NoOngoingSale {});
-    };
-
-    // Check if sale has started
-    ensure!(
-        sale.start_time.is_expired(&ctx.env.block),
-        ContractError::SaleNotStarted {}
-    );
-    // Check if sale has ended
-    if let Some(end_time) = sale.end_time {
-        ensure!(
-            !end_time.is_expired(&ctx.env.block),
-            ContractError::SaleEnded {}
-        );
-    }
-
-    let purchased = amount_sent.checked_div(sale.exchange_rate).unwrap();
-    let remainder = amount_sent.checked_sub(purchased.checked_mul(sale.exchange_rate)?)?;
-
-    ensure!(
-        !purchased.is_zero(),
-        ContractError::InvalidFunds {
-            msg: "Not enough funds sent to purchase a token".to_string()
-        }
-    );
-    ensure!(sale.amount >= purchased, ContractError::NotEnoughTokens {});
-
-    // If purchase was rounded down return funds to purchaser
-    if !remainder.is_zero() {
-        resp = resp
-            .add_submessage(generate_transfer_message(
-                asset_sent.clone(),
-                remainder,
-                sender.to_string(),
-                Some(REFUND_REPLY_ID),
-            )?)
-            .add_attribute("refunded_amount", remainder);
-    }
-
-    // Transfer tokens to purchaser recipient
-    let token_addr = TOKEN_ADDRESS
-        .load(deps.storage)?
-        .get_raw_address(&deps.as_ref())?;
-
-    let token_asset = AssetInfo::Cw20(token_addr);
-    let sub_msg = generate_transfer_message_recipient(
-        &deps.as_ref(),
-        token_asset,
-        purchased,
-        recipient.clone(),
-        Some(PURCHASE_REPLY_ID),
-    )?;
-
-    resp = resp.add_submessage(sub_msg);
-
-    // Update sale amount remaining
-    sale.amount = sale.amount.checked_sub(purchased)?;
-    SALE.save(deps.storage, &asset_sent.inner(), &sale)?;
-
-    // Transfer exchanged asset to recipient
-    resp = resp.add_submessage(generate_transfer_message_recipient(
-        &deps.as_ref(),
-        asset_sent.clone(),
-        amount_sent - remainder,
-        sale.recipient.clone(),
-        Some(RECIPIENT_REPLY_ID),
-    )?);
-
-    Ok(resp.add_attributes(vec![
-        attr("action", "purchase"),
-        attr("purchaser", sender),
-        attr("recipient", recipient.address.to_string()),
-        attr("amount", purchased),
-        attr("purchase_asset", asset_sent.to_string()),
-        attr("purchase_asset_amount_send", amount_sent - remainder),
-        attr("recipient", sale.recipient.address.to_string()),
-    ]))
-}
-
-pub fn execute_redeem(
-    ctx: ExecuteContext,
-    amount_sent: Uint128,
-    asset_sent: AssetInfo,
-    recipient: Recipient,
-    // For refund purposes
-    sender: &str,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { deps, .. } = ctx;
-
-    let mut resp = Response::default();
-
-    let Some(mut redeem) = REDEEM.may_load(deps.storage, &asset_sent.inner())? else {
-        return Err(ContractError::NoOngoingRedeem {});
-    };
-
-    // Check if redeem has started
-    ensure!(
-        redeem.start_time.is_expired(&ctx.env.block),
-        ContractError::RedeemNotStarted {}
-    );
-    // Check if sale has ended
-    if let Some(end_time) = redeem.end_time {
-        ensure!(
-            !end_time.is_expired(&ctx.env.block),
-            ContractError::RedeemEnded {}
-        );
-    }
-
-    let payment_decimal = Decimal::from_ratio(amount_sent, 1u128);
-    let tokens_to_receive_decimal = payment_decimal.checked_mul(redeem.exchange_rate)?;
-    let potential_redeemed = tokens_to_receive_decimal.to_uint_floor();
-
-    // Calculate actual redemption amounts
-    let (redeemed_amount, amount_received, refund_amount) = if potential_redeemed <= redeem.amount {
-        (potential_redeemed, amount_sent, Uint128::zero())
-    } else {
-        // If we don't have enough tokens, calculate the partial redemption
-        let actual_redeemed = redeem.amount;
-
-        // Convert to Decimal for calculation
-        let redeem_amount_decimal = Decimal::from_ratio(redeem.amount, 1u128);
-        let actual_amount_needed_decimal = redeem_amount_decimal
-            .checked_div(redeem.exchange_rate)
-            .map_err(|_| ContractError::Overflow {})?;
-        let actual_amount_needed = actual_amount_needed_decimal.to_uint_ceil();
-
-        let refund = amount_sent.checked_sub(actual_amount_needed)?;
-        (actual_redeemed, actual_amount_needed, refund)
-    };
-
-    ensure!(
-        !redeemed_amount.is_zero(),
-        ContractError::InvalidFunds {
-            msg: "Not enough funds sent to purchase a token".to_string()
-        }
-    );
-    ensure!(
-        redeem.amount >= redeemed_amount,
-        ContractError::NotEnoughTokens {}
-    );
-
-    // If purchase was rounded down return funds to purchaser
-    if !refund_amount.is_zero() {
-        resp = resp
-            .add_submessage(generate_transfer_message(
-                asset_sent.clone(),
-                refund_amount,
-                sender.to_string(),
-                Some(REFUND_REPLY_ID),
-            )?)
-            .add_attribute("refunded_amount", refund_amount);
-    }
-
-    // Transfer tokens to the user that's redeeming
-    let redeem_asset = redeem.asset.clone();
-    let redeem_recipient = redeem.clone().recipient;
-
-    let transfer_msg = generate_transfer_message_recipient(
-        &deps.as_ref(),
-        redeem_asset.clone(),
-        redeemed_amount,
-        recipient.clone(),
-        None,
-    )?;
-    resp = resp.add_submessage(transfer_msg);
-
-    // Update redeem amount remaining
-    redeem.amount = redeem.amount.checked_sub(redeemed_amount)?;
-    redeem.amount_paid_out = redeem.amount_paid_out.checked_add(redeemed_amount)?;
-    REDEEM.save(deps.storage, &asset_sent.inner(), &redeem)?;
-
-    // Transfer exchanged asset to recipient
-    resp = resp.add_submessage(generate_transfer_message_recipient(
-        &deps.as_ref(),
-        asset_sent.clone(),
-        amount_sent - refund_amount,
-        redeem_recipient.clone(),
-        Some(RECIPIENT_REPLY_ID),
-    )?);
-
-    Ok(resp.add_attributes(vec![
-        attr("action", "redeem"),
-        attr("redeemer", sender),
-        attr("recipient", recipient.address.to_string()),
-        attr("amount", amount_received),
-        attr("redeem_asset", redeem_asset.to_string()),
-        attr("redeem_asset_amount_send", amount_sent - refund_amount),
-        attr("recipient", redeem_recipient.address.to_string()),
-    ]))
-}
-
-pub fn execute_purchase_native(
-    ctx: ExecuteContext,
-    recipient: Option<Recipient>,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { ref info, .. } = ctx;
-    let sender = info.sender.to_string();
-    let recipient = Recipient::validate_or_default(recipient, &ctx, sender.as_str())?;
-
-    // Only allow one coin for purchasing
-    let payment = one_coin(info)?;
-    let asset = AssetInfo::Native(payment.denom.to_string());
-    let amount = payment.amount;
-
-    execute_purchase(ctx, amount, asset, recipient, &sender)
-}
-
-pub fn execute_start_redeem_native(
-    ctx: ExecuteContext,
-    redeem_asset: AssetInfo,
-    exchange_rate: Decimal,
-    recipient: Option<Recipient>,
-    start_time: Option<Expiry>,
-    end_time: Option<Milliseconds>,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { ref info, .. } = ctx;
-
-    let native_funds_sent = one_coin(info)?;
-    let amount_sent = native_funds_sent.amount;
-    let asset_sent = AssetInfo::Native(native_funds_sent.denom.to_string());
-    let sender = info.sender.to_string();
-
-    execute_start_redeem(
-        ctx,
-        amount_sent,
-        asset_sent,
-        redeem_asset,
-        exchange_rate,
-        sender,
-        recipient,
-        start_time,
-        end_time,
-    )
-}
-
-pub fn execute_redeem_native(
-    ctx: ExecuteContext,
-    recipient: Option<Recipient>,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { ref info, .. } = ctx;
-
-    let sender = info.sender.to_string();
-    let asset_sent = one_coin(info)?;
-    let amount_sent = asset_sent.amount;
-    let asset_sent = AssetInfo::Native(asset_sent.denom.to_string());
-
-    let recipient = Recipient::validate_or_default(recipient, &ctx, sender.as_str())?;
-
-    execute_redeem(ctx, amount_sent, asset_sent, recipient, &sender)
-}
-
-pub fn execute_cancel_sale(
-    ctx: ExecuteContext,
-    asset: AssetInfo,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { deps, info, .. } = ctx;
-
-    let Some(sale) = SALE.may_load(deps.storage, &asset.to_string())? else {
-        return Err(ContractError::NoOngoingSale {});
-    };
-
-    let mut resp = Response::default();
-
-    // Refund any remaining amount
-    if !sale.amount.is_zero() {
-        let token_addr = TOKEN_ADDRESS
-            .load(deps.storage)?
-            .get_raw_address(&deps.as_ref())?;
-
-        let token = AssetInfo::Cw20(token_addr);
-        resp = resp
-            .add_submessage(generate_transfer_message(
-                token,
-                sale.amount,
-                info.sender.to_string(),
-                Some(REFUND_REPLY_ID),
-            )?)
-            .add_attribute("refunded_amount", sale.amount);
-    }
-
-    // Sale can now be removed
-    SALE.remove(deps.storage, &asset.to_string());
-
-    Ok(resp.add_attributes(vec![
-        attr("action", "cancel_sale"),
-        attr("asset", asset.to_string()),
-    ]))
-}
-
-pub fn execute_cancel_redeem(
-    ctx: ExecuteContext,
-    asset: AssetInfo,
-) -> Result<Response, ContractError> {
-    let ExecuteContext { deps, info, .. } = ctx;
-
-    let Some(redeem) = REDEEM.may_load(deps.storage, &asset.inner())? else {
-        return Err(ContractError::NoOngoingRedeem {});
-    };
-
-    let mut resp = Response::default();
-
-    // Refund any remaining amount
-    if !redeem.amount.is_zero() {
-        let token = redeem.asset;
-        resp = resp
-            .add_submessage(generate_transfer_message(
-                token,
-                redeem.amount,
-                info.sender.to_string(),
-                Some(REFUND_REPLY_ID),
-            )?)
-            .add_attribute("refunded_amount", redeem.amount);
-    }
-
-    // Sale can now be removed
-    SALE.remove(deps.storage, &asset.to_string());
-
-    Ok(resp.add_attributes(vec![
-        attr("action", "cancel_redeem"),
-        attr("asset", asset.to_string()),
-    ]))
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
